--- command/doc.go
@@ -1,6 +1,0 @@
-/*Package command provides methods and functions for creating and parsing Altid control messages
-
-	go get github.com/altid/server/command
-
-*/
-package command
