+++ internal/routes/errors.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/altid/server/files"
+	"github.com/altid/server/internal/message"
 )
 
 // Nothing needs to be synthesized here
@@ -14,12 +14,12 @@ type ErrHandler struct{}
 
 func NewError() *ErrHandler { return &ErrHandler{} }
 
-func (*ErrHandler) Normal(msg *files.Message) (interface{}, error) {
+func (*ErrHandler) Normal(msg *message.Message) (interface{}, error) {
 	fp := path.Join(msg.Service, "errors")
 	return os.Open(fp)
 }
 
-func (*ErrHandler) Stat(msg *files.Message) (os.FileInfo, error) {
+func (*ErrHandler) Stat(msg *message.Message) (os.FileInfo, error) {
 	fp := path.Join(msg.Service, "errors")
 	return os.Lstat(fp)
 }
