+++ internal/routes/normal.go
@@ -4,19 +4,19 @@ import (
 	"os"
 	"path"
 
-	"github.com/altid/server/files"
+	"github.com/altid/server/internal/message"
 )
 
 type NormalHandler struct{}
 
 func NewNormal() *NormalHandler { return &NormalHandler{} }
 
-func (*NormalHandler) Normal(msg *files.Message) (interface{}, error) {
+func (*NormalHandler) Normal(msg *message.Message) (interface{}, error) {
 	fp := path.Join(msg.Service, msg.Buffer, msg.Target)
 	return os.Open(fp)
 }
 
-func (*NormalHandler) Stat(msg *files.Message) (os.FileInfo, error) {
+func (*NormalHandler) Stat(msg *message.Message) (os.FileInfo, error) {
 	fp := path.Join(msg.Service, msg.Buffer, msg.Target)
 	return os.Lstat(fp)
 }
