+++ internal/tabs/doc.go
@@ -1,0 +1,7 @@
+/*Package tabs contains helper functions for managing tabs in Altid services
+
+	go get github.com/altid/server/tabs
+
+
+*/
+package tabs
