+++ internal/tabs/resources/tabs
@@ -1,0 +1,5 @@
+Item one
+Item two
+ThirdThing
+item4
+5th_element
