--- tabs/resources/tabs
@@ -1,5 +1,0 @@
-Item one
-Item two
-ThirdThing
-item4
-5th_element
